kedro-ibis-poc/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kedro-ibis-poc"
version = "0.1.0"
description = "Kedro+Ibis POC with source-faker integration"
requires-python = ">=3.10,<3.12"
dependencies = [
    "kedro~=0.19.0",
    "kedro-datasets[ibis-duckdb]~=4.0",
    "airbyte~=0.24.2",
    "ibis-framework[duckdb]~=9.0",
    "duckdb~=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "basedpyright>=1.28.2",
    "pytest>=8.3.5",
]

[tool.kedro]
package_name = "kedro_ibis_poc"
project_name = "Kedro Ibis POC"
kedro_init_version = "0.19.0"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
